Dozvoli dodavanje u korpu samo kad je knjiga dostupna

+++ E-library/Services/ShoppingCart/ShoppingCart.API/Controllers/CartController.cs
@@ -16,7 +16,7 @@ using ShoppingCart.API.Services;
 
 namespace ShoppingCart.API.Controllers
 {
-    //[Authorize(Roles="Member")]
+    [Authorize(Roles="Member")]
     [ApiController]
     [Route("api/v1/[controller]")]
     public class CartController : ControllerBase
@@ -105,6 +105,16 @@ namespace ShoppingCart.API.Controllers
         [ProducesResponseType(typeof(Cart), StatusCodes.Status200OK)]
         public async Task<ActionResult<Cart>> AddBookToCart(string username, string bookId)
         {
+            if (User.FindFirst(ClaimTypes.Name).Value != username)
+            {
+                return Forbid();
+            }
+
+            if (_service.AddBookToCart(username, bookId) == null)
+            {
+                return Forbid();
+            }
+
             return Ok(await _service.AddBookToCart(username, bookId));
         }
 
@@ -113,6 +123,11 @@ namespace ShoppingCart.API.Controllers
         [ProducesResponseType(typeof(Cart), StatusCodes.Status200OK)]
         public async Task<ActionResult<Cart>> RemoveBookFromCart(string username, string bookId)
         {
+            if (User.FindFirst(ClaimTypes.Name).Value != username)
+            {
+                return Forbid();
+            }
+
             return Ok(await _service.RemoveBookFromCart(username, bookId));
         }
